requirements.txt
openai~=1.14.3
aiogram~=3.16.0
pytesseract~=0.3.10
pillow~=11.0.0
python-dotenv~=1.0.1
pytest~=8.3.4
psycopg2-binary~=2.9.9
aiohttp~=3.9.1
pypdf2~=3.0.1
python-docx~=1.1.0
celery~=5.3.6